move headers to common.h
